Add in Basic Server Provider Tests
Just the basics that get past 500 errors.  We can start fleshing out the
bodies and checking the data is correctly filled in next.
